+++ public/components/visualize/wz-visualize.js
@@ -19,6 +19,7 @@ import {
   EuiFlexGroup,
   EuiPanel,
   EuiFlexItem,
+  EuiButton,
   EuiButtonIcon,
   EuiDescriptionList,
   EuiCallOut,
@@ -34,6 +35,7 @@ import { Metrics } from '../overview/metrics/metrics';
 import { PatternHandler } from '../../react-services/pattern-handler';
 import { getToasts } from '../../kibana-services';
 import { SecurityAlerts } from './components';
+import { toMountPoint } from '../../../../../src/plugins/kibana_react/public';
 
 const visHandler = new VisHandlers();
 
@@ -58,6 +60,7 @@ export class WzVisualize extends Component {
     this.monitoringEnabled = !!(configuration || {})[
       'wazuh.monitoring.enabled'
     ];
+    this.newFields={};
   }
 
 
@@ -126,14 +129,17 @@ export class WzVisualize extends Component {
     this.setState({ expandedVis: this.state.expandedVis === id ? false : id });
   };
 
-  refreshKnownFields = async () => {
+  refreshKnownFields = async ( newField = null ) => {
+    if(newField && newField.name){
+      this.newFields[newField.name] = newField;
+    }
     if (!this.state.hasRefreshedKnownFields) { // Known fields are refreshed only once per dashboard loading
       try {
         this.setState({ hasRefreshedKnownFields: true, isRefreshing: true });
-        await PatternHandler.refreshIndexPattern();
+        await PatternHandler.refreshIndexPattern(this.newFields);
         this.setState({ isRefreshing: false });
-        this.showToast('success', 'The index pattern was refreshed successfully.');
-
+        this.reloadToast();
+        this.newFields={};
       } catch (err) {
         this.setState({ isRefreshing: false });
         this.showToast('danger', 'The index pattern could not be refreshed');
@@ -144,7 +150,21 @@ export class WzVisualize extends Component {
       await this.refreshKnownFields();
     }
   }
-
+  reloadToast = () => {
+    getToasts().add({
+      color: 'success',
+      title: 'The index pattern was refreshed successfully.',
+      text: toMountPoint(<EuiFlexGroup justifyContent="flexEnd" gutterSize="s">
+        <EuiFlexItem grow={false}>
+          There were some unknown fields for the current index pattern.
+          You need to refresh the page to apply the changes.
+        </EuiFlexItem>
+        <EuiFlexItem grow={false}>
+          <EuiButton onClick={() => window.location.reload()} size="s">Reload page</EuiButton>
+        </EuiFlexItem>
+      </EuiFlexGroup>)
+    })
+  }
   render() {
     const { visualizations } = this.state;
     const { selectedTab } = this.props;