+++ public/kibana-integrations/kibana-vis.js
@@ -33,6 +33,7 @@ import {
   EuiFlexGroup,
 } from "@elastic/eui";
 import { getAngularModule, getToasts, getVisualizationsPlugin, getSavedObjects, getDataPlugin, getChrome, getOverlays } from '../kibana-services';
+import { KnownFields } from "../utils/known-fields";
 
 class KibanaVis extends Component {
   _isMounted = false;
@@ -57,7 +58,7 @@ class KibanaVis extends Component {
     this.tabVisualizations = new TabVisualizations();
     this.state = {
       visRefreshingIndex: false,
-    };   
+    };
     const services = {
       savedObjectsClient: getSavedObjects().client,
       indexPatterns: getDataPlugin().indexPatterns,
@@ -127,7 +128,7 @@ class KibanaVis extends Component {
           data.value.visData &&
           data.value.visData.rows &&
           this.props.state[this.visID] !==
-            data.value.visData.rows["0"]["col-0-1"]
+          data.value.visData.rows["0"]["col-0-1"]
         ) {
           store.dispatch(
             this.updateMetric({
@@ -148,7 +149,7 @@ class KibanaVis extends Component {
           data.value.visData.tables["0"].rows &&
           data.value.visData.tables["0"].rows["0"] &&
           this.props.state[this.visID] !==
-            data.value.visData.tables["0"].rows["0"]["col-0-2"]
+          data.value.visData.tables["0"].rows["0"]["col-0-2"]
         ) {
           store.dispatch(
             this.updateMetric({
@@ -370,7 +371,7 @@ class KibanaVis extends Component {
           this.visualization.searchSource.setField("source", false);
           // Visualization doesn't need "hits"
           this.visualization.searchSource.setField('size', 0);
-          const visState = await getVisualizationsPlugin().convertToSerializedVis(this.visualization);        
+          const visState = await getVisualizationsPlugin().convertToSerializedVis(this.visualization);
           const vis = await getVisualizationsPlugin().createVis(this.visualization.visState.type, visState);
           this.visHandler = await getVisualizationsPlugin().__LEGACY.createVisEmbeddableFromObject(
             vis,
@@ -415,7 +416,11 @@ class KibanaVis extends Component {
           this.visualization = null;
           this.renderInProgress = false;
           this.rendered = false;
-          await this.props.refreshKnownFields();
+
+          // if there's a field name it looks for known fields structures          
+          const foundField = (match[1] && KnownFields.find(field => field.name === match[1].trim()));
+          
+          await this.props.refreshKnownFields(foundField);
         }
         this.renderInProgress = false;
         return this.myRender(raw);
@@ -472,11 +477,11 @@ class KibanaVis extends Component {
   destroyAll = () => {
     try {
       this.visualization.destroy();
-    } catch (error) {} // eslint-disable-line
+    } catch (error) { } // eslint-disable-line
     try {
       this.visHandler.destroy();
       this.visHandler = null;
-    } catch (error) {} // eslint-disable-line
+    } catch (error) { } // eslint-disable-line
   };
 
   renderComplete = async () => {
@@ -532,7 +537,7 @@ class KibanaVis extends Component {
               paddingTop: 100,
             }}
           >
-            <EuiFlexGroup style={{placeItems: 'center'}}>
+            <EuiFlexGroup style={{ placeItems: 'center' }}>
               <EuiFlexItem>
               </EuiFlexItem>
               <EuiFlexItem grow={false}>
@@ -560,8 +565,8 @@ class KibanaVis extends Component {
             style={{
               display:
                 this.deadField &&
-                this.props.resultState !== "loading" &&
-                !this.state.visRefreshingIndex
+                  this.props.resultState !== "loading" &&
+                  !this.state.visRefreshingIndex
                   ? "block"
                   : "none",
               textAlign: "center",